% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vowelchangerfun.R
\name{vowel_changer}
\alias{vowel_changer}
\title{Title}
\usage{
vowel_changer(character_string, type_change = "remove", capitalization = FALSE)
}
\arguments{
\item{character_string}{a character string that will be changed}

\item{type_change}{options are "remove" or "replace" to remove all vowels or replace all vowels}

\item{capitalization}{options are TRUE/FALSE. False keeps as is, true capitalizes the result}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{            A string that came in will be changed slightly and printed
}\if{html}{\out{</div>}}
}
\description{
Title
}
\examples{
vowel_changer(character_string = "The quick red fox jumped over the lazy brown dog", type_change = "remove", capitalization = FALSE)

}
